Use strncmp() instead of strcmp() in util_storeData
util_storeData would only intermittantly miss a tag, so with a safe compare it
shouldn't have to worry about garbage on the end of the tag.  So far it seems to
be working well.

Before creating util_storeData id3_storeData used strncmp as well, and there was
probably a reason for that.